requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
dashscope==1.13.0
moviepy==1.0.3
alibabacloud-alimt20181012==1.0.0
alibabacloud-tea-openapi==0.3.8
alibabacloud-tea-util==0.3.11
playsound==1.3.0
requests==2.31.0
werkzeug==2.3.7 